backend/tools_deployment/covid-cxr-cloudrun/requirements.txt
flask
gunicorn
pillow
numpy
tensorflow-cpu==2.15.0